requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
google==3.0.0
google-api-core==2.8.1
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-cloud-bigquery==3.18.0
google-cloud-bigquery-storage==2.13.1
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
googlemaps==4.10.0
python-dotenv==1.0.0
python-multipart==0.0.6
pyautogen==0.2.34
pytz==2024.2
requests==2.31.0
bs4==0.0.1
image==1.5.33
matplotlib==3.9.2
matplotlib-inline==0.1.6
seaborn==0.12.2
numpy==1.26.4
pandas==2.2.2
scipy==1.14.1
openai==1.56.0
tiktoken==0.5.2
firecrawl-py==0.0.20
# twine==5.1.1
# twilio==9.2.3
tavily-python==0.4.0
fonttools==4.53.1
markdown2==2.5.0
fpdf==1.7.2
fpdf2==2.7.9
firebase-admin==6.0.1
db-dtypes==1.3.0
pydantic==2.10.2
# staticmap==0.5.7
# geopy==2.4.1
openpyxl==3.1.5
requests==2.31.0
requests-oauthlib==2.0.0
bs4==0.0.1
python-docx==1.1.2